requirements.txt
flask
torch
torchvision
ultralytics
roboflow
opencv-python
pillow
numpy
werkzeug
easyocr
